% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{CopulaSim}
\alias{CopulaSim}
\title{CopulaSim}
\usage{
CopulaSim(n, R, list_qlaws, repetition, random = FALSE, vartype)
}
\arguments{
\item{n}{Integer corresponding to the number of observations.}

\item{list_qlaws}{List, list of  quasi-inverse functions (must be chosen among (qbeta, qbinom, qchisq, qubif, qexp, qf, qgamma, qgeom, qhyper, qnbinom, qnorm, qpois, qt, qweibull)) with their specified parameters}

\item{repetition}{vector, number of variables following each law}

\item{random}{if TRUE the probability distributions will be randomly distributed among the variables, if FALSE their order will be preserved. (by default : random = TRUE)}

\item{vartype}{a vector containing the letters "C" (continuous) and "D"(discrete) denoting the type of each variable, in order}

\item{blocks}{vector which contains the size of the blocks on the diagonal of the matrix}

\item{coeff}{vector of coefficients associated to each block (there is no need to mention the "1" coefficient for one dimensional blocks)}
}
\value{
a list containing a matrix of d vectors (in columns) of length n such that the joint distribution function of those vectors is given  by F(.,.) = C(F_1(.),..., F_d(.)) where C is the gaussian Copula related to a block diagonal matrix and F_1,.., F_d are respectively the distribution functions of the margins, and a vector containing the re-ordered variable types
}
\description{
This function enables the user to simulate a dataset of n observations by d variables which cumulative distribution function corresponds to a Gaussian copula. The simulation is done via the generalized inverse method.
}
\examples{
M <- diag_block_matrix(c(3,2,4), c(0.2,0.1,0.5))
CopulaSim(100, M, list(function(p){qexp(p=p, rate=1)}, function(p){qnorm(p=p, mean=0, sd=1)}, function(p){qnorm(p=p, mean=0.76, sd=2)}), c(3,4,2), random = TRUE)

}
