% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{cor_network_graph}
\alias{cor_network_graph}
\title{cor_network_graph}
\usage{
cor_network_graph(
  data,
  Type,
  TS,
  parallel = TRUE,
  Weighted = FALSE,
  color,
  legend
)
}
\arguments{
\item{data}{the dataset containing d variables (columns) on n observations}

\item{Type}{a vector containing the type of each variable or the group it belongs to}

\item{TS}{scalar, defined the threshold at which the random variables are considered as being correlated}

\item{Weighted}{boolean, if TRUE, edges will be weighted (the smaller the estimated coefficient the longer the edge). (Default to FALSE)}

\item{color}{a vector containing colors for the vertices}

\item{legend}{a vector containing the legend}
}
\value{
a graph where the vertices are correlations and the edges are random variables.
}
\description{
Create a correlation network for heterogeneous data
}
\examples{
data = CopulaSim(100, c(3,2,4), c(0.5,0.8,0.3), list(function(p){qexp(p=p, rate=1)}, function(p){qnorm(p=p, mean=0, sd=1)}, function(p){qnorm(p=p, mean=0.76, sd=2)}), c(5,2,2), random = TRUE,vartype=rep("C",9))
cor_network_graph(data[[1]],data[[2]],0.3, TRUE)

}
