% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{diag_block_matrix}
\alias{diag_block_matrix}
\title{diag_block_matrix}
\usage{
diag_block_matrix(blocks, coeff)
}
\arguments{
\item{blocks}{a vector containing the size of each block}

\item{coeff}{a vector containing the coefficient corresponding to each block}
}
\value{
the generated block matrix
}
\description{
This function enables the user to generate a diagonal matrix of size d x d
}
\examples{
M <- diag_block_matrix(c(3,1,4), c(0.2,0.1,0.5))
}
