% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{fdr_d}
\alias{fdr_d}
\title{fdr_d}
\usage{
fdr_d(X, Type)
}
\arguments{
\item{X}{data frame of d variables and n observations}

\item{Type}{a vector of length d containing the type of each variable, "C" denoting continuous, "D" denoting discrete}
}
\value{
an array of size 2 x n x d where the first dimension contains the n x d values for F- (0 for continuous variables), the second one contains the n x d values for F
}
\description{
This function enables the user to estimate the empirical cumulative distribution function of each variable
}
\examples{
M <- diag_block_matrix(c(3,2,4), c(0.2,0.1,0.5))
 data <- CopulaSim(100, M, list(function(p){qexp(p=p, rate=1)}, function(p){qnorm(p=p, mean=0, sd=1)}, function(p){qnorm(p=p, mean=0.76, sd=2)}), c(3,4,2), random = TRUE)
 fdr_d(data[[1]],data[[2]])

}
