% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{GCopula}
\alias{GCopula}
\title{GCopula}
\usage{
GCopula(R, n)
}
\arguments{
\item{R}{the correlation matrix of the copula. It has to be a dxd positive semi-definite matrix.}

\item{n}{the number of observations}
}
\value{
The dataset, a data frame containing d vectors (in columns) of length n containing the values F_1(X_1), ..., F_d(X_d) such that the joint distribution function of those vectors is given by F(.,.) = C(F_1(.),..., F_d(.)) where C is the Gaussian Copula of correlation matrix R. F_1,.., F_d denote the marginal distribution functions of the variables X_1, ..., X_d.
}
\description{
This function enables the user to simulate a dataset of n observations by d variables which cumulative distribution function corresponds to a multivariate Gaussian cumulative distribution function of correlation matrix R. In the context of our Gaussian copula model, it corresponds to the simulation of our marginals F_1(X_1), ..., F_d(X_d).
}
\examples{
M <- diag_block_matrix(c(3,1,4), c(0.2,0.1,0.5))
 GCopula(M, 100)
}
