% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_XY.R
\name{list_XY}
\alias{list_XY}
\title{list_XY}
\usage{
list_XY(R, n, list_qlaws)
}
\arguments{
\item{R}{Matrix, 2x2 correlation matrix}

\item{n}{Integer, length of the simulated data}

\item{list_qlaws}{List, list of two quasi-inverse functions of one argument}
}
\value{
list of two vectors of length n such that the joint distribution function of those vectors is given  by F(.,.) = C(F_1(.), F_2(.)) where C is the gaussian Copula related to R and F_1, F_2 are respectively the distribution functions of the margins.
}
\description{
2D data simulator based on the generalized inverse method
}
\examples{
list_XY(matrix(c(1,0.5,0.5,1),2,2), 100, list(loi1 = function(p){qexp(p=p, rate=1)}, loi2 = function(p){qnorm(p=p, mean=0, sd=1)}))
}
