% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_XY_d.R
\name{list_XY_d}
\alias{list_XY_d}
\title{Title list_XY_d}
\usage{
list_XY_d(n, blocks, coeff, list_qlaws, list_param, repetition, random = TRUE)
}
\arguments{
\item{n}{Integer, length of the simulated data}

\item{blocks}{vector, vector with the size of the blocks on the diagonal of the matrix}

\item{coeff}{vector, coefficient associated to each block (there is no need to inform the coefficient for one dimensional block (see examples))}

\item{list_qlaws}{List, list of  quasi-inverse functions (must be chosen among (qbeta, qbinom, qchisq, qubif, qexp, qf, qgamma, qgeom, qhyper, qnbinom, qnorm, qpois, qt, qweibull))}

\item{list_param}{List, list with lists of parameters}

\item{repetition}{vector, number of repetition of each laws}

\item{random}{if TRUE the distribution of each laws will be randomly distibuted among the blocks. if FALSE the order enter as a parameter will be preserved. (by default : random = TRUE)}
}
\value{
list of d vectors of length n such that the joint distribution function of those vectors is given  by F(.,.) = C(F_1(.),..., F_d(.)) where C is the gaussian Copula related to a block diagonal matrix and F_1,.., F_d are respectively the distribution functions of the margins.
}
\description{
Title list_XY_d
}
\examples{
list_XY_d(100, c(3,1,4,5), c(0.5,0.8,0.3), list(qexp, qnorm, qbinom), list(3, c(0,1), c(20,0.5)), c(5,4,4), random = TRUE)
}
