% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{matrix_cor_ts}
\alias{matrix_cor_ts}
\title{matrix_cor_ts}
\usage{
matrix_cor_ts(M_est, TS, binary = TRUE)
}
\arguments{
\item{M_est}{a correlation matrix}

\item{TS}{a threshold}

\item{binary}{a boolean specifying whether the coefficients should be binarized, TRUE by defaut}
}
\value{
the thresholded input matrix
}
\description{
This function enables the user to threshold matrix coefficients
}
\examples{
M <- diag_block_matrix(c(3,4,5),c(0.7,0.8,0.2))
matrix_cor_ts(M,0.5)

}
