% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{matrix_gen}
\alias{matrix_gen}
\title{matrix_gen}
\usage{
matrix_gen(d, gamma)
}
\arguments{
\item{d}{the number of columns/rows of the matrix}

\item{gamma}{initial sparsity parameter}
}
\value{
a list containing the generated matrix and its new sparsity parameter (ie the proportion of zeros)
}
\description{
This function enables the user to generate a positive semi-definite sparse matrix of size d x d and with initial sparsity parameter gamma.
}
\examples{
matrix_gen(30,0.2)
}
