% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{rho_estim}
\alias{rho_estim}
\title{rho_estim}
\usage{
rho_estim(data, Type, parallel = TRUE)
}
\arguments{
\item{data}{data frame containing the random vectors in columns}

\item{Type}{a vector of characters corresponding to the type of each function, "C" for continuous and "D" for discrete.}

\item{parallel}{TRUE if you want the code to be parallelized, FALSE if not}
}
\value{
A correlation matrix with the estimates of the correlation coefficients of the copula.
}
\description{
given d random vectors X_1, .., X_d, this function estimates the correlation matrix associated to the gaussian copula C such that the joint distribution function of the random vectors is C(F1(.), ..., Fd(.)) with F1, ..., Fd the cumulative distribution functions of the margins.
}
\examples{
data = CopulaSim(100, c(3,2,4), c(0.5,0.8,0.3), list(function(p){qexp(p=p, rate=1)}, function(p){qnorm(p=p, mean=0, sd=1)}, function(p){qnorm(p=p, mean=0.76, sd=2)}), c(3,4,2), random = TRUE,rep("C",9))
          rho_estim(data[[1]], data[[2]])

}
