% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_estim_2.R
\name{rho_estim_2}
\alias{rho_estim_2}
\title{rho_estim_2}
\usage{
rho_estim_2(X, Y)
}
\arguments{
\item{X}{list, list of iid random variables}

\item{Y}{list, list of iid random variables (with the same length as X)}
}
\value{
an estimate of the correlation matrix parameter
}
\description{
rho_estim_2
}
