% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_estim_d.R
\name{rho_estim_d}
\alias{rho_estim_d}
\title{rho_estim_d}
\usage{
rho_estim_d(list_var, Shape)
}
\arguments{
\item{list_var}{list, list of d random vectors with the same length}

\item{Shape}{The shape returned by the function ("vector" or "matrix")}
}
\value{
A correlation matrix or a vector with the estimates of the correlation coefficients
}
\description{
given d random vectors, this function estimates the correlations matrix associated to the gaussian copula C such that the joint distribution function of the random vectors is C(F1(.), ..., Fd(.)) with F1, ..., Fd the distribution functions of the margins.
}
\examples{
list_X = list_XY_d(100, c(3,1,4,5), c(0.5,0.8,0.3), list(qexp, qnorm, qbinom), list(3, c(0,1), c(20,0.5)), c(5,4,4), random = TRUE)
          rho_estim_d(list_X, "matrix")
}
