% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{CopulaSim}
\alias{CopulaSim}
\title{CopulaSim}
\usage{
CopulaSim(n, R, qdist, rep, random = FALSE)
}
\arguments{
\item{n}{the number of observations}

\item{R}{a correlation matrix of size dxd}

\item{qdist}{a list containing the marginal quantile functions}

\item{rep}{a vector containing the number of replicates of each distribution}

\item{random}{a boolean defining whether the correlation matrix should be shuffled}
}
\value{
a list containing an nxd data frame, the shuffled correlation matrix R, and the shuffling order
}
\description{
This function enables the user to simulate data which joint cumulative distribution function can be expressed as a Gaussian copula
}
\examples{
M <- diag_block_matrix(c(3,4,5),c(0.7,0.8,0.2))
CopulaSim(20,M,c("qnorm(0,1)","qexp(0.5)","qbinom(4,0.8)"),c(6,4,2),random=TRUE)

}
