% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_network_graph.R
\name{cor_network_graph}
\alias{cor_network_graph}
\title{cor_network_graph}
\usage{
cor_network_graph(
  list_var,
  TS,
  Weighted = FALSE,
  vertex_col = "lightsteelblue2"
)
}
\arguments{
\item{list_var}{list, list of d random vectors with the same length}

\item{TS}{scalar, defined the threshold at which the random variables are considered as being correlated}

\item{Weighted}{boolean, if TRUE, edges will be weighted (the smaller the estimated coefficient the longer the edge). (Default to FALSE)}

\item{vertex_col}{a color for the vertices of the graph (default color : ligth blue)}
}
\value{
a graph where the vertices are correlations and the edges are random variables.
}
\description{
cor_network_graph
}
\examples{
cor_network_graph_ (list(X,Y,Z), 0.3, TRUE, "orange")

}
