% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterocop.R
\name{cor_network_graph}
\alias{cor_network_graph}
\title{cor_network_graph}
\usage{
cor_network_graph(data, R, TS, binary = TRUE, legend)
}
\arguments{
\item{R}{a correlation matrix of size dxd}

\item{TS}{a threshold for the coefficients of the correlation matrix}

\item{binary}{a boolean specifying whether the coefficients should be binarized, TRUE by defaut. If FALSE, the edge width is proportional to the values of coefficients.}

\item{legend}{a vector containing the type of each variable used to color the vertices}

\item{n}{the number of observations}
}
\value{
a graph representing the correlations between the variables
}
\description{
This function enables the user to plot the graph corresponding to the correlations of the Gaussian copula
}
\examples{
R <- diag_block_matrix(c(3,4,5),c(0.7,0.8,0.2))
data <- CopulaSim(20,R,c("qnorm(0,1)","qexp(0.5)","qbinom(4,0.8)"),c(6,4,2),random=FALSE)[[1]]
cor_network_graph(data,R,TS=0.3,binary=TRUE,legend=c(rep("Normal",6),rep("Exponential",4),rep("Binomial",2)))

}
