% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_XY.R
\name{list_XY}
\alias{list_XY}
\title{Data generator}
\usage{
list_XY(R, n, list_qlaws, list_param)
}
\arguments{
\item{R}{Matrix, 2x2 correlation matrix}

\item{n}{Integer, length of the simulated data}

\item{list_qlaws}{List, list of two quasi-inverse function ( must be chosen among (qbeta, qbinom, qchisq, qubif, qexp, qf, qgamma, qgeom, qhyper, qnbinom, qnorm, qpois, qt, qweibull))}

\item{list_param}{List, list with lists of parameters}
}
\value{
list of two vectors of length n such that the joint distribution function of those vectors is given  by F(.,.) = C(F_1(.), F_2(.)) where C is the gaussian Copula related to R and F_1, F_2 are respectively the distribution functions of the margins.
}
\description{
Data generator
}
\examples{
list_XY(matrix(c(1,0.5,0.5,1),2,2), 100, list(loi1 = qexp, loi2 = qnorm), list(3,c(0,1)))
}
