% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_estim_d.R
\name{rho_estim_d}
\alias{rho_estim_d}
\title{rho_estim_d}
\usage{
rho_estim_d(list_var, Shape)
}
\arguments{
\item{list_var}{list, list of d random vectors with the same length}

\item{Shape}{The shape returned by the function ("vector" or "matrix")}
}
\value{
A correlation matrix or a vector with the estimates of the correlation coefficients
}
\description{
rho_estim_d
}
\examples{
rho_estim_d(list(X,Y,Z), "matrix")
}
